pyproject.toml
[tool.poetry]
name = "user"
version = "0.1.0"
description = ""
authors = ["Gabriel Luís Fernando Vieira de Souza <gabriel-soz@live.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
hiredis = "^2.0.0"
aioredis = "^2.0.1"
orjson = "^3.7.2"

[tool.poetry.dev-dependencies]
commitizen = "^2.27.1"
pydantic-factories = "^1.3.0"
faker-vehicle = "^0.2.0"
boto3 = "^1.24.12"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
